BCGlobalToken·BCSemanticToken 확장을 별도 파일로 분리
생성 파일(BCGlobalToken.swift, BCSemanticToken.swift)에 섞여 있던 수기 확장을
`{Type}+{Topic}.swift` 파일로 분리한다. 파일명은 같은 Foundation 계층의 기존
관례(BTSemanticToken+Style.swift, BTSemanticToken+Typography.swift)를 따랐다.

- BCGlobalToken+Color.swift: color / uiColor
- BCSemanticToken+SemanticColorProtocol.swift: SemanticColorProtocol 준수
- BCSemanticToken+PressedColor.swift: pressedColor

paletteSet의 private을 제거해 internal로 완화했다. Swift의 private은 같은 파일
내에서만 접근 가능하므로, light/dark를 다른 파일로 옮기면 private인 채로는
컴파일되지 않는다. 함께 선언된 typealias PaletteSet이 이미 internal이라 접근
수준은 일관된다. 이에 맞춰 더 이상 사실이 아닌 MARK 주석도 갱신했다.
